エンコードはいったん手元に書き出してからNASに格納する

+++ app/controllers/queue.rb
@@ -79,8 +79,11 @@ Frogbit.controllers :queue do
     end
     EM.defer do
       result = encode_queue.encode()
-      encode_queue.video.update(:is_encoded=>true, :encode_log=>result[:log], :saved_directory=>$config[:output_dir])
-      encode_queue.destroy
+
+      if result.result
+        encode_queue.video.update(:is_encoded=>true, :encode_log=>result[:log], :saved_directory=>$config[:output_dir])
+        encode_queue.destroy
+      end
     end
 
     flash[:info] = "「#{encode_queue.video.output_name}」のエンコードを開始しました"
